init: add packages for dev
