+++ .gitignore
@@ -1,5 +1,6 @@
 # dependencies
 /node_modules
+/back
 
 # System Files
 .DS_Store
