+++ package.json
@@ -1,21 +1,23 @@
 {
-    "name": "ws_task",
-    "version": "1.0.0",
-    "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start:dev": "npx nodemon",
+        "build": "tsc",
+        "start:prod": "tsc && node back/index.js",
+        "lint": "eslint . --ext .ts"
     },
-    "author": "Andrei Auchynnikau",
-    "keywords": [],
-    "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/node": "^20.11.17",
+        "@typescript-eslint/eslint-plugin": "^7.0.1",
+        "@typescript-eslint/parser": "^7.0.1",
+        "eslint": "^8.56.0",
+        "eslint-config-prettier": "^9.1.0",
+        "nodemon": "^3.0.3",
+        "prettier": "^3.2.5",
+        "ts-node": "^10.9.2",
+        "ts-node-dev": "^2.0.0",
+        "typescript": "^5.3.3"
     },
     "dependencies": {
-        "ws": "^8.8.0"
+        "ws": "^8.16.0"
     }
 }
